remove create device component
