+++ console/console-init/ui/src/components/RootModal/RootModal.tsx
@@ -12,7 +12,6 @@ import {
 } from "modules/address-space/dialogs";
 import { EditAddress, CreateAddress } from "modules/address/dialogs";
 import { UpdatePassword } from "components/UpdatePassword";
-import { CreateDevice } from "modules/iot-device/dialogs";
 
 const MODAL_COMPONENTS: any = {
   [MODAL_TYPES.CREATE_ADDRESS_SPACE]: CreateMessagingProject,
@@ -27,8 +26,7 @@ const MODAL_COMPONENTS: any = {
   [MODAL_TYPES.UPDATE_PASSWORD]: UpdatePassword,
   [MODAL_TYPES.DELETE_PROJECT]: DialogPrompt,
   [MODAL_TYPES.DELETE_IOT_DEVICE]: DialogPrompt,
-  [MODAL_TYPES.UPDATE_DEVICE_STATUS]: DialogPrompt,
-  [MODAL_TYPES.CREATE_DEVICE]: CreateDevice
+  [MODAL_TYPES.UPDATE_DEVICE_STATUS]: DialogPrompt
 };
 
 export const RootModal: React.FC<{}> = () => {
