+++ console/console-init/ui/src/context-state-reducer/actions.ts
@@ -41,7 +41,6 @@ const DELETE_IOT_DEVICE = "DELETE_IOT_DEVICE";
 const SET_DEVICE_ACTION_TYPE = "SET_DEVICE_ACTION_TYPE";
 const RESET_DEVICE_ACTION_TYPE = "RESET_DEVICE_ACTION_TYPE";
 const UPDATE_DEVICE_STATUS = "UPDATE_DEVICE_STATUS";
-const CREATE_DEVICE = "CREATE_DEVICE";
 
 const types = {
   SET_SERVER_ERROR,
@@ -68,8 +67,7 @@ const MODAL_TYPES = {
   UPDATE_PASSWORD,
   DELETE_PROJECT,
   DELETE_IOT_DEVICE,
-  UPDATE_DEVICE_STATUS,
-  CREATE_DEVICE
+  UPDATE_DEVICE_STATUS
 };
 
 export { MODAL_TYPES, types };
