+++ console/console-init/ui/src/modules/iot-device/dialogs/CreateDevice/index.ts
@@ -3,5 +3,4 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export * from "./CreateDevice";
 export * from "./DeviceInformation";
