--- console/console-init/ui/src/stories/CreateDevice.stories.tsx
@@ -1,21 +1,0 @@
-import React from "react";
-import { MemoryRouter } from "react-router";
-import { CreateDevice } from "modules/iot-device/dialogs";
-import { action } from "@storybook/addon-actions";
-
-export default {
-  title: "Create Device"
-};
-
-export const createDeviceWizard = () => {
-  return (
-    <MemoryRouter>
-      <CreateDevice
-        setPropertyInput={action("property set")}
-        onPropertyClear={action("property clear clicked")}
-        onPropertySelect={action("property select clicked")}
-        onChangePropertyInput={async () => {}}
-      />
-    </MemoryRouter>
-  );
-};
